Removes variable scaling in loadvar
Previously scaling and offsetting of netCDF variables was applied
amnually, but this is no longer necessary on lxlp, so it has been
commented for now.

In general this is probably a question of whether the cdo bindings
are using netCDF4 or not. A check should be introduced to make
sure the scaling is done if required, but not otherwise.